requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# PyTorch ecosystem
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# PyTorch Geometric
torch-geometric>=2.4.0
torch-geometric-temporal>=0.54.0

# Time series
pmdarima>=2.0.0
statsmodels>=0.14.0

# Data processing
pyarrow>=14.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
folium>=0.15.0

# Web application
streamlit>=1.29.0
streamlit-folium>=0.17.0

# API clients
kaggle>=1.5.16

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Logging
structlog>=23.2.0

# Development
black>=23.12.0
isort>=5.13.0
mypy>=1.7.0
ruff>=0.1.8

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
